requirements.txt
python-dotenv==1.0.0
librosa==0.10.1
numba==0.58.1
numpy>=1.24.0
sounddevice==0.4.6
soundfile==0.12.1
scikit-learn==1.3.2
whisper==1.1.10